Move the unknown solver constant to the front of the list
Previously when it was 99, if the status were printed with an unknown
solver the printing would index outside the solver name array and print
garbage instead of unknown. The value at index 0 was previously unused,
and there was actually a printing error previously (since it indexed an
array using these values and the array started indexing at 0).