Fix Windows CI to use latest image and VS version
GitHub is in the process of rolling out the windows-2022 image as
windows-latest, so some runners may use that. The CMake generator we
need has changed though (since the VS version was updated), meaning our
CI is only compatible with one image version. Hardcode the newest image
to ensure our CMake generator works.