Prefix linear solver constants with OSQP to prevent clashes (#395)
* Prefix linear solver constants with OSQP to prevent clashes

* Move the unknown solver constant to the front of the list

Previously when it was 99, if the status were printed with an unknown
solver the printing would index outside the solver name array and print
garbage instead of unknown. The value at index 0 was previously unused,
and there was actually a printing error previously (since it indexed an
array using these values and the array started indexing at 0).

* Fix Windows CI to use latest image and VS version

GitHub is in the process of rolling out the windows-2022 image as
windows-latest, so some runners may use that. The CMake generator we
need has changed though (since the VS version was updated), meaning our
CI is only compatible with one image version. Hardcode the newest image
to ensure our CMake generator works.