codegen: Write out nnull vector variables as defines
We can't directly assign a null value to a pointer we use in a struct
because then it ends up throwing an "initializer element is not
constant" error when compiling the struct. Instead define the pointer to
just be a preprocessor define to the null pointer, so the struct gets
initialized directly with a null pointer.