Allow csc copy routine to copy an empty matrix
Previously, the copy routine would allocate an x and i vector no matter
the input vector. This meant that if the vector passed in was empty (no
values) the resulting copy of the vector would have allocated vectors,
but no values to put in the vectors.

This would cause a problem with codegen because it would then reference
uninitialized memory when writing out an empty constraint matrix.