Simplify implementation of MKL's vector plus/minus
These can be simple one-line calls to the VML functions for aadd and
subtract instead of requiring conditionals and copies.

Simplify the API for the OSQPVectorf_minus function to remove the
special case on the b arguement. This wasn't actually implemented in the
built-in algebra, and was never used in the actual calling code.