Add another comment about fall through in the sample
GCC is stupid when using -Wextra and refuses to accept any explanation
for why the fall through is occuring in the actual comment string that
will silence the compiler error. Instead, it requires the comment to be
all alone and not-informative as to WHY we want to fall through, so we
have to include two comments that say the same thing to make it not warn
anymore.