Add support for GPU arrays to be passed in/taken out of OSQPVectorf/i
This adds support for OSQPVectorf/i to copy data directly from/to arrays
allocated on the GPU as well as the host. This means users can pass data
already on the GPU into OSQP, as well as ask for OSQP to put it onto the
GPU when done (instead of having to go through the host for all memory
transfers).