Consolidate dy_l and dy_u into a single dy for adjoint derivatives
The need for both lower and upper bound perturbations of the dual
variables is an implementation detail of the current implicit
differentiation method, and does not always need to be the case.
Instead, the user-facing API should just have a single dy input and the
differentiation method should then handle the duplication internally.